chore: convert to ts
